requirements.txt
en-core-web-sm==3.0.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
fasttext==0.9.2
ftfy==6.0.3
future==0.18.2
futures==3.1.1
gast==0.4.0
gmpy2==2.0.8
googletrans==3.0.0
goslate==1.5.1
graphviz==0.15
h5py==3.1.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.12.22
json-model==1.0.1
json5==0.9.5
jupyter==1.0.0
Keras==2.4.3
keras-models==0.0.7
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
langdetect==1.0.9
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas-profiling==3.0.0
parso==0.8.2
pathlib==1.0.1
pipreqs==0.4.10
pkginfo==1.7.0
plotly==4.14.3
python-language-server==0.31.7
python3-utils==0.4.0
QtPy==1.9.0
rake-nltk==1.0.4
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.2
sentencepiece==0.1.96
shellingham==1.3.1
simplegeneric==0.8.1
sklearn==0.0
spyder==4.0.1
spyder-kernels==1.8.1
statsmodels==0.12.2
stopwords==1.0.0
tables==3.6.1
tangled-up-in-unicode==0.1.0
tb-nightly==2.6.0a20210626
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.9.4
textblob==0.15.3
tf-estimator-nightly==2.6.0.dev2021062501
tf-nightly==2.7.0.dev20210626
tflearn==0.5.0
tokenization==1.0.7
unicodecsv==0.14.1
utils==1.0.1
visions==0.7.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wordcount==1.0
